Divide project into individual components
