+++ fastapi_hypermodel/__init__.py
@@ -1,10 +1,16 @@
-from .hal import HALFor, HALForType, HalHyperModel, HALResponse
-from .hypermodel import (
+from .base import (
+    URL_TYPE_SCHEMA,
     AbstractHyperField,
     HasName,
     HyperModel,
+    InvalidAttribute,
+    UrlType,
+    extract_value_by_name,
+    get_hal_link_href,
+    get_route_from_app,
+    resolve_param_values,
 )
-from .linkset import LinkSet, LinkSetType
+from .hal import HALFor, HALForType, HalHyperModel, HALLinkType, HALResponse
 from .siren import (
     SirenActionFor,
     SirenActionType,
@@ -18,27 +24,18 @@ from .siren import (
     get_siren_link,
 )
 from .url_for import UrlFor
-from .url_type import URL_TYPE_SCHEMA, UrlType
-from .utils import (
-    InvalidAttribute,
-    extract_value_by_name,
-    get_hal_link_href,
-    get_route_from_app,
-    resolve_param_values,
-)
 
 __all__ = [
     "URL_TYPE_SCHEMA",
     "AbstractHyperField",
     "HALFor",
     "HALForType",
+    "HALLinkType",
     "HALResponse",
     "HalHyperModel",
     "HasName",
     "HyperModel",
     "InvalidAttribute",
-    "LinkSet",
-    "LinkSetType",
     "SirenActionFor",
     "SirenActionType",
     "SirenEmbeddedType",
