+++ fastapi_hypermodel/base/url_type.py
