+++ fastapi_hypermodel/base/utils.py
