+++ fastapi_hypermodel/siren/siren_schema.py
