+++ fastapi_hypermodel/url_for/__init__.py
@@ -1,0 +1,3 @@
+from .url_for import UrlFor, UrlForType
+
+__all__ = ["UrlFor", "UrlForType"]
